feat: move the functions to utils
